refactor: moved icon visualization from the marker service to the map store
